Add files via upload
I am having issues with the text wrap exercise. I am able to write the code to recreate the example, but every time I try to run my code for the user input function by jupyter wont run it and times out. I think it's close, but likely missing a parentheses or something.